requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
alembic
